refactor : add DrugName enum, move classes to model
