+++ index.js
@@ -1,12 +1,13 @@
-import { Drug, Pharmacy } from "./pharmacy";
-
+import { Pharmacy } from "./models/pharmacy";
+import { DrugName } from "./models/DrugName";
+import { Drug } from "./models/Drug";
 import fs from "fs";
 
 const drugs = [
-  new Drug("Doliprane", 20, 30),
-  new Drug("Herbal Tea", 10, 5),
-  new Drug("Fervex", 5, 40),
-  new Drug("Magic Pill", 15, 40)
+  new Drug(DrugName.DOLIPRANE, 20, 30),
+  new Drug(DrugName.HERBAL_TEA, 10, 5),
+  new Drug(DrugName.FERVEX, 5, 40),
+  new Drug(DrugName.MAGIC_PILL, 15, 40)
 ];
 const trial = new Pharmacy(drugs);
