+++ models/DrugName.js
@@ -1,0 +1,6 @@
+export const DrugName = {
+  DOLIPRANE: "Doliprane",
+  HERBAL_TEA: "Herbal Tea",
+  FERVEX: "Fervex",
+  MAGIC_PILL: "Magic Pill"
+};
