+++ models/Pharmacy.js
@@ -1,10 +1,4 @@
-export class Drug {
-  constructor(name, expiresIn, benefit) {
-    this.name = name;
-    this.expiresIn = expiresIn;
-    this.benefit = benefit;
-  }
-}
+import { DrugName } from "./DrugName";
 
 export class Pharmacy {
   constructor(drugs = []) {
@@ -13,18 +7,18 @@ export class Pharmacy {
   updateBenefitValue() {
     for (var i = 0; i < this.drugs.length; i++) {
       if (
-        this.drugs[i].name != "Herbal Tea" &&
-        this.drugs[i].name != "Fervex"
+        this.drugs[i].name != DrugName.HERBAL_TEA &&
+        this.drugs[i].name != DrugName.FERVEX
       ) {
         if (this.drugs[i].benefit > 0) {
-          if (this.drugs[i].name != "Magic Pill") {
+          if (this.drugs[i].name != DrugName.MAGIC_PILL) {
             this.drugs[i].benefit = this.drugs[i].benefit - 1;
           }
         }
       } else {
         if (this.drugs[i].benefit < 50) {
           this.drugs[i].benefit = this.drugs[i].benefit + 1;
-          if (this.drugs[i].name == "Fervex") {
+          if (this.drugs[i].name == DrugName.FERVEX) {
             if (this.drugs[i].expiresIn < 11) {
               if (this.drugs[i].benefit < 50) {
                 this.drugs[i].benefit = this.drugs[i].benefit + 1;
@@ -38,14 +32,14 @@ export class Pharmacy {
           }
         }
       }
-      if (this.drugs[i].name != "Magic Pill") {
+      if (this.drugs[i].name != DrugName.MAGIC_PILL) {
         this.drugs[i].expiresIn = this.drugs[i].expiresIn - 1;
       }
       if (this.drugs[i].expiresIn < 0) {
-        if (this.drugs[i].name != "Herbal Tea") {
-          if (this.drugs[i].name != "Fervex") {
+        if (this.drugs[i].name != DrugName.HERBAL_TEA) {
+          if (this.drugs[i].name != DrugName.FERVEX) {
             if (this.drugs[i].benefit > 0) {
-              if (this.drugs[i].name != "Magic Pill") {
+              if (this.drugs[i].name != DrugName.MAGIC_PILL) {
                 this.drugs[i].benefit = this.drugs[i].benefit - 1;
               }
             }
